+++ vinfo/CHANGELOG.rst
@@ -26,3 +26,9 @@ This changelog includes all changes to the cog since 06/03/2021.
 
 * Add dependencies in aliases and usage
 * Use trigger typing when getting attributes
+
+----------
+20/03/2021
+----------
+
+* Fixes for sys, math, array and more, when they did not have ``__file__`` attributes.